refactor: Python 源码按 core/chat/render/webui 分包
入口 main.py 与 constants.py 留在根；pages/ 扫描路径不变。
同步修正包内相对导入与插件根路径（assets、_conf_schema、metadata）。
