+++ chat/flavor_profiles.py
