+++ core/hapi_client.py
