+++ core/hapi_routes.py
