+++ core/notification_manager.py
