+++ render/__init__.py
@@ -1,0 +1,1 @@
+"""Package namespace for hapi connector."""
